support admin page, and support --url_prefix
