+++ pyproject.toml
@@ -24,4 +24,4 @@ build-backend = "poetry.core.masonry.api"
 line-length = 79
 
 [tool.poetry.scripts]
-prometheus-http-sd = 'prometheus_http_sd.app:main'
+prometheus-http-sd = 'prometheus_http_sd.cli:main'
